created a useInpputError hook
